* Made changes as per code review
