+++ mlfinlab/labeling/labeling.py
@@ -9,36 +9,6 @@ import pandas as pd
 from mlfinlab.util.multiprocess import mp_pandas_obj
 
 
-def get_daily_vol(close, lookback=100):
-    """
-    Snippet 3.1, page 44, Daily Volatility Estimates
-
-    Computes the daily volatility at intraday estimation points.
-
-    In practice we want to set profit taking and stop-loss limits that are a function of the risks involved
-    in a bet. Otherwise, sometimes we will be aiming too high (tao ≫ sigma_t_i,0), and sometimes too low
-    (tao ≪ sigma_t_i,0 ), considering the prevailing volatility. Snippet 3.1 computes the daily volatility
-    at intraday estimation points, applying a span of lookback days to an exponentially weighted moving
-    standard deviation.
-
-    See the pandas documentation for details on the pandas.Series.ewm function.
-
-    Note: This function is used to compute dynamic thresholds for profit taking and stop loss limits.
-
-    :param close: Closing prices
-    :param lookback: lookback period to compute volatility
-    :return: series of daily volatility value
-    """
-    # daily vol re-indexed to close
-    df0 = close.index.searchsorted(close.index - pd.Timedelta(days=1))
-    df0 = df0[df0 > 0]
-    df0 = (pd.Series(close.index[df0 - 1], index=close.index[close.shape[0] - df0.shape[0]:]))
-
-    df0 = close.loc[df0.index] / close.loc[df0.values].values - 1  # daily returns
-    df0 = df0.ewm(span=lookback).std()
-    return df0
-
-
 # Snippet 3.2, page 45, Triple Barrier Labeling Method
 def apply_pt_sl_on_t1(close, events, pt_sl, molecule):  # pragma: no cover
     """
@@ -50,7 +20,7 @@ def apply_pt_sl_on_t1(close, events, pt_sl, molecule):  # pragma: no cover
     Mainly it returns a DataFrame of timestamps regarding the time when the first barriers were reached.
 
     :param close: (series) close prices
-    :param events: (series) of indices that signify "events" (see get_t_events function
+    :param events: (series) of indices that signify "events" (see cusum_filter function
     for more details)
     :param pt_sl: (array) element 0, indicates the profit taking level; element 1 is stop loss level
     :param molecule: (an array) a set of datetime index values for processing
@@ -115,7 +85,8 @@ def add_vertical_barrier(t_events, close, num_days=1):
 
 
 # Snippet 3.3 -> 3.6 page 50, Getting the Time of the First Touch, with Meta Labels
-def get_events(close, t_events, pt_sl, target, min_ret, num_threads, vertical_barrier_times=False, side_prediction=None):
+def get_events(close, t_events, pt_sl, target, min_ret, num_threads, vertical_barrier_times=False,
+               side_prediction=None):
     """
     Snippet 3.6 page 50, Getting the Time of the First Touch, with Meta Labels
 
@@ -158,8 +129,8 @@ def get_events(close, t_events, pt_sl, target, min_ret, num_threads, vertical_ba
         side_ = side_prediction.loc[target.index]
         pt_sl_ = pt_sl[:2]
 
-    events = pd.concat({'t1': vertical_barrier_times, 'trgt': target, 'side': side_},
-                       axis=1)
+    # Create a new df with [v_barrier, target, side] and drop rows that are NA in target
+    events = pd.concat({'t1': vertical_barrier_times, 'trgt': target, 'side': side_}, axis=1)
     events = events.dropna(subset=['trgt'])
 
     # Apply Triple Barrier
